refactor(config): use uint8 for backlog workers and queue size

+++ config.h
@@ -3,9 +3,9 @@
 extern char *address;
 extern uint16_t port;
 
-extern int backlog;
-extern int workers;
-extern int queue_size;
+extern uint8_t backlog;
+extern uint8_t workers;
+extern uint8_t queue_size;
 
 extern int bwt_ttl;
 extern char *bwt_key;
